selenium tests for bidding
